+++ VendingMachine/src/DTO/Item.java
@@ -11,10 +11,10 @@ package DTO;
  */
 public class Item {
     private String name;// Item name
-    private float cost;// Item cost
+    private double cost;// Item cost
     private int left;// Number of items in inventory
     
-    public Item(String name, float cost, int left){
+    public Item(String name, double cost, int left){
         this.name = name;
         this.cost = cost;
         this.left = left;
@@ -24,8 +24,8 @@ public class Item {
     public String getName(){ return this.name; }
     public void setName(String name) { this.name = name; }
     
-    public float getCost(){ return this.cost; }
-    public void setCost(float cost) { this.cost = cost; }
+    public double getCost(){ return this.cost; }
+    public void setCost(double cost) { this.cost = cost; }
     
     public int getLeft(){ return this.left; }
     public void setLeft(int left) { this.left = left; }
